improve load_area functionality and error report
In area_config.load_area, add the functionality to load from a
pathlib.Path object or a stream.  Deprecate the functionality to load
from a string directly --- temporarily, it will assume a string
descirbes a yaml file if it contains a newline, and a path otherwise.
If a file does not exist or is not readable, leave it to the open
function to raise the appropriate exception.

Closes #250.