Adapt to satisfy deepcode.
I disagree with deepcodes suggestion.  Explicit is better than implicit.
PEP 8 does NOT recommend against ``if len(x) == 0``, it recommends
against ``if not len(x)``.  According to "Explicit is better than
implicit", ``if len(x) == 0`` should be preferable and clearer.
However, I will adapt it anyway as deepcode is the boss here?