Prealloc appendable array by segment count & size
The appendable array class now pre-allocates memory lazily based on
expected number of segments and the size of the first segment received.
This way we don't need to add another pre-allocation size parameter to
the interface, and still make sure that no unnecessary RAM is occupied.
For instance, when users utilize masked arrays for neighbour info
queries index and distance arrays can be significantly smaller, as is
demonstrated in the `test_nearest_masked_swath_target`test.