Reserve number of elements instead of segments
This seems to cover the most common cases where we expect the final
array to be roughly in the ballpark of the target area's size.
However, since it still can happen that pixels turn out to be invalid,
we need to trim at the end. Also, we initialise lazily, so we know the
full shape of the expected segments.